LMS: Add GetPlatformServiceRecordRaw to credentials registry
Add GetPlatformServiceRecordRaw to the credentials
processed by LMS in registry.

+++ UNS/AMT_COM_Interface/AMT_COM_Interface.cpp
@@ -114,6 +114,8 @@ HRESULT AddRegKeys()
 		return 1;
 	if (false == ds.SetDataValue(GetPlatformServiceRecord_F, L"Administrators", true))
 		return 1;
+	if (false == ds.SetDataValue(GetPlatformServiceRecordRaw_F, L"Administrators", true))
+		return 1;
 
 	if (false == ds.DeleteDataVal(NETWORK_TRAFFIC_TX_CEASED_))
 		return 1;
